ADD: Use cases for get all added

+++ src/main/java/com/hardwareStore/hardwareStore/model/ProductDTO.java
@@ -9,7 +9,7 @@ public class ProductDTO {
     private String id;
     private String name;
     private String description;
-    private Double price;
+    private Integer price;
     private Integer minUnits;
     private Integer maxUnits;
     private Integer quantity;
